Restructured crowd member resetting to avoid errors
Now, ResetRoute is not a coroutine. Instead, it just stops and doesn't do anything else if there are no positions or if the chosen start is on cooldown.

Now CrowdMembers themselves repeatedly request to be reset, until they are reset, every second (not every frame). This avoids an issue with destroyed buildings being removed from the arrays CrowdManager uses.